feat(conf): replace $DIR by the dir. of the model file (llama-server flags)
When using models like GPT OSS, we need to provide a grammar file.
see: https://github.com/ggml-org/llama.cpp/discussions/15396#discussioncomment-14145537
We want to have the possibility to keep the model and grammar files within the same directory.
But we also want to be free to move that directory
without having to update the path within tho command line arguments.
Thus, we use $DIR as a placeholder for the directory.